Extract RecordsCollectionReference from GroundFirestore
